move OTM tests to separate class + make cleanup part of context config
