+++ molgenis-platform-integration-tests/src/test/java/org/molgenis/integrationtest/platform/PlatformIT.java
@@ -1,13 +1,10 @@
 package org.molgenis.integrationtest.platform;
 
-import com.google.common.collect.Lists;
 import org.apache.commons.io.FileUtils;
 import org.molgenis.data.*;
 import org.molgenis.data.cache.l2.L2Cache;
-import org.molgenis.data.cache.l3.L3Cache;
 import org.molgenis.data.elasticsearch.SearchService;
 import org.molgenis.data.elasticsearch.index.job.IndexService;
-import org.molgenis.data.i18n.I18nUtils;
 import org.molgenis.data.i18n.LanguageService;
 import org.molgenis.data.i18n.model.I18nStringMetaData;
 import org.molgenis.data.i18n.model.LanguageFactory;
@@ -20,14 +17,7 @@ import org.molgenis.data.support.DynamicEntity;
 import org.molgenis.data.support.QueryImpl;
 import org.molgenis.test.data.EntitySelfXrefTestHarness;
 import org.molgenis.test.data.EntityTestHarness;
-import org.molgenis.test.data.OneToManyTestHarness;
 import org.molgenis.test.data.staticentity.TestEntityStatic;
-import org.molgenis.test.data.staticentity.bidirectional.person1.PersonMetaData1;
-import org.molgenis.test.data.staticentity.bidirectional.person2.PersonMetaData2;
-import org.molgenis.test.data.staticentity.bidirectional.person3.PersonMetaData3;
-import org.molgenis.test.data.staticentity.bidirectional.person4.PersonMetaData4;
-import org.molgenis.test.data.staticentity.bidirectional.authorbook1.AuthorMetaData1;
-import org.molgenis.test.data.staticentity.bidirectional.authorbook1.BookMetaData1;
 import org.molgenis.util.EntityUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -39,13 +29,19 @@ import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.testng.AbstractTestNGSpringContextTests;
-import org.springframework.transaction.annotation.Transactional;
-import org.testng.annotations.*;
+import org.testng.annotations.AfterMethod;
+import org.testng.annotations.BeforeClass;
+import org.testng.annotations.DataProvider;
+import org.testng.annotations.Test;
 
+import javax.annotation.PreDestroy;
 import java.io.File;
 import java.io.IOException;
 import java.text.ParseException;
-import java.util.*;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Locale;
+import java.util.Set;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Supplier;
@@ -53,12 +49,10 @@ import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
 
 import static com.google.common.collect.Lists.newArrayList;
-import static com.google.common.collect.Sets.newHashSet;
 import static java.util.Arrays.asList;
 import static java.util.Collections.emptyList;
 import static java.util.Collections.singletonList;
 import static java.util.stream.Collectors.toList;
-import static java.util.stream.Collectors.toSet;
 import static java.util.stream.Stream.concat;
 import static java.util.stream.Stream.of;
 import static org.molgenis.data.RepositoryCapability.*;
@@ -69,7 +63,6 @@ import static org.molgenis.data.meta.model.EntityMetaData.AttributeCopyMode.DEEP
 import static org.molgenis.data.meta.model.EntityMetaDataMetaData.ENTITY_META_DATA;
 import static org.molgenis.security.core.runas.RunAsSystemProxy.runAsSystem;
 import static org.molgenis.test.data.EntityTestHarness.*;
-import static org.molgenis.test.data.OneToManyTestHarness.*;
 import static org.molgenis.util.MolgenisDateFormat.getDateFormat;
 import static org.molgenis.util.MolgenisDateFormat.getDateTimeFormat;
 import static org.testng.Assert.*;
@@ -78,7 +71,6 @@ import static org.testng.Assert.*;
 public class PlatformIT extends AbstractTestNGSpringContextTests
 {
 	private final Logger LOG = LoggerFactory.getLogger(PlatformIT.class);
-	private final String GROUPS_ONE_TO_MANY = "oneToMany";
 
 	private EntityMetaData entityMetaDataStatic;
 	private EntityMetaData refEntityMetaDataStatic;
@@ -92,8 +84,6 @@ public class PlatformIT extends AbstractTestNGSpringContextTests
 	private EntityTestHarness testHarness;
 	@Autowired
 	private EntitySelfXrefTestHarness entitySelfXrefTestHarness;
-	@Autowired
-	private OneToManyTestHarness oneToManyTestHarness;
 	@Autowired
 	private DataService dataService;
 	@Autowired
@@ -119,13 +109,7 @@ public class PlatformIT extends AbstractTestNGSpringContextTests
 	@Autowired
 	private LanguageFactory languageFactory;
 	@Autowired
-	private EntityMetaDataFactory entityMetaDataFactory;
-	@Autowired
 	private AttributeMetaDataFactory attributeMetaDataFactory;
-	@Autowired
-	private EntityManager entityManager;
-	@Autowired
-	private L3Cache l3Cache;
 
 	/**
 	 * Wait till the whole index is stable. Index job is done a-synchronized.
@@ -187,7 +171,7 @@ public class PlatformIT extends AbstractTestNGSpringContextTests
 		waitForWorkToBeFinished(indexService, LOG);
 	}
 
-	private List<GrantedAuthority> makeAuthorities(String entityName, boolean write, boolean read, boolean count)
+	static List<GrantedAuthority> makeAuthorities(String entityName, boolean write, boolean read, boolean count)
 	{
 		List<GrantedAuthority> authorities = newArrayList();
 		if (write) authorities.add(new SimpleGrantedAuthority("ROLE_ENTITY_WRITE_" + entityName.toUpperCase()));
@@ -213,60 +197,8 @@ public class PlatformIT extends AbstractTestNGSpringContextTests
 		authorities.addAll(makeAuthorities(i18nStringMetaData.getName(), true, false, false));
 		authorities.addAll(makeAuthorities(entityMetaDataMetaData.getName(), true, true, true));
 
-		for (int i = 1; i <= ONE_TO_MANY_CASES; i++)
-		{
-			authorities.addAll(makeAuthorities("sys_Author" + i, true, true, true));
-			authorities.addAll(makeAuthorities("sys_Book" + i, true, true, true));
-			authorities.addAll(makeAuthorities("sys_Person" + i, true, true, true));
-		}
-
 		SecurityContextHolder.getContext()
 				.setAuthentication(new TestingAuthenticationToken("user", "user", authorities));
-	}
-
-	@AfterClass
-	public void cleanUp() throws InterruptedException
-	{
-		// Give asyncTransactionLog time to stop gracefully
-		TimeUnit.SECONDS.sleep(1);
-
-		applicationContext.close();
-		SecurityContextHolder.getContext().setAuthentication(null);
-
-		try
-		{
-			// Delete molgenis home folder
-			FileUtils.deleteDirectory(new File(System.getProperty("molgenis.home")));
-		}
-		catch (IOException e)
-		{
-			LOG.error("Error removing molgenis home directory", e);
-		}
-		l2Cache.logStatistics();
-		cleanUpLanguages();
-	}
-
-	private void cleanUpLanguages()
-	{
-		List<AttributeMetaData> languageAttrs = new ArrayList<>();
-		for (AttributeMetaData attr : attributeMetaDataMetaData.getAttributes())
-		{
-			if (I18nUtils.isI18n(attr.getName()))
-			{
-				languageAttrs.add(attr);
-			}
-		}
-		languageAttrs.forEach(attributeMetaDataMetaData::removeAttribute);
-
-		languageAttrs.clear();
-		for (AttributeMetaData attr : entityMetaDataMetaData.getAttributes())
-		{
-			if (I18nUtils.isI18n(attr.getName()))
-			{
-				languageAttrs.add(attr);
-			}
-		}
-		languageAttrs.forEach(entityMetaDataMetaData::removeAttribute);
 	}
 
 	@AfterMethod
@@ -285,23 +217,6 @@ public class PlatformIT extends AbstractTestNGSpringContextTests
 		waitForIndexToBeStable(entityMetaDataDynamic.getName(), indexService, LOG);
 		waitForIndexToBeStable(refEntityMetaDataDynamic.getName(), indexService, LOG);
 		waitForIndexToBeStable(selfXrefEntityMetaData.getName(), indexService, LOG);
-	}
-
-	@AfterMethod(groups = GROUPS_ONE_TO_MANY)
-	public void afterGroupsOneToMany()
-	{
-		runAsSystem(() ->
-		{
-			deleteAuthorsThenBooks(1);
-			deleteBooksThenAuthors(2);
-			deleteAuthorsThenBooks(3);
-			deleteAuthorsThenBooks(4);
-			dataService.deleteAll(PersonMetaData1.NAME);
-			dataService.deleteAll(PersonMetaData2.NAME);
-			dataService.deleteAll(PersonMetaData3.NAME);
-			dataService.deleteAll(PersonMetaData4.NAME);
-		});
-		waitForWorkToBeFinished(indexService, LOG);
 	}
 
 	private void addDefaultLanguages()
@@ -1392,393 +1307,6 @@ public class PlatformIT extends AbstractTestNGSpringContextTests
 		q6.search("searchTestBatchUpdate");
 		Stream<Entity> result6 = searchService.searchAsStream(q6, entityMetaDataDynamic);
 		assertEquals(result6.count(), 1);
-	}
-
-	@Test(singleThreaded = true, groups = GROUPS_ONE_TO_MANY, dataProvider = "oneToManyTestCaseDataProvider")
-	public void testOneToManyAuthorAndBookInsert(int testCase)
-	{
-		OneToManyTestHarness.AuthorsAndBooks authorsAndBooks = importAuthorsAndBooks(testCase);
-
-		String bookEntityName = authorsAndBooks.getBookMetaData().getName();
-		assertEquals(dataService.findOneById(bookEntityName, BOOK_1).getEntity(ATTR_AUTHOR).getIdValue(), AUTHOR_1);
-		assertEquals(dataService.findOneById(bookEntityName, BOOK_2).getEntity(ATTR_AUTHOR).getIdValue(), AUTHOR_2);
-		assertEquals(dataService.findOneById(bookEntityName, BOOK_3).getEntity(ATTR_AUTHOR).getIdValue(), AUTHOR_3);
-
-		String authorEntityName = authorsAndBooks.getAuthorMetaData().getName();
-		assertEquals(dataService.findOneById(authorEntityName, AUTHOR_1).getEntities(ATTR_BOOKS).iterator().next()
-				.getIdValue(), BOOK_1);
-		assertEquals(dataService.findOneById(authorEntityName, AUTHOR_2).getEntities(ATTR_BOOKS).iterator().next()
-				.getIdValue(), BOOK_2);
-		assertEquals(dataService.findOneById(authorEntityName, AUTHOR_3).getEntities(ATTR_BOOKS).iterator().next()
-				.getIdValue(), BOOK_3);
-	}
-
-	@Test(singleThreaded = true, groups = GROUPS_ONE_TO_MANY, dataProvider = "oneToManyTestCaseDataProvider")
-	public void testOneToManyPersonInsert(int testCase)
-	{
-		List<Entity> persons = importPersons(testCase);
-
-		String personEntityName = persons.get(0).getEntityMetaData().getName();
-		Entity person1 = dataService.findOneById(personEntityName, PERSON_1);
-		Entity person2 = dataService.findOneById(personEntityName, PERSON_2);
-		Entity person3 = dataService.findOneById(personEntityName, PERSON_3);
-
-		assertEquals(person1.getEntity(ATTR_PARENT).getIdValue(), PERSON_3);
-		assertEquals(person2.getEntity(ATTR_PARENT).getIdValue(), PERSON_1);
-		assertEquals(person3.getEntity(ATTR_PARENT).getIdValue(), PERSON_2);
-
-		assertEquals(
-				StreamSupport.stream(person1.getEntities(ATTR_CHILDREN).spliterator(), false).map(Entity::getIdValue)
-						.collect(toSet()), newHashSet(PERSON_2));
-		assertEquals(
-				StreamSupport.stream(person2.getEntities(ATTR_CHILDREN).spliterator(), false).map(Entity::getIdValue)
-						.collect(toSet()), newHashSet(PERSON_3));
-		assertEquals(
-				StreamSupport.stream(person3.getEntities(ATTR_CHILDREN).spliterator(), false).map(Entity::getIdValue)
-						.collect(toSet()), newHashSet(PERSON_1));
-	}
-
-	@DataProvider(name = "oneToManyTestCaseDataProvider")
-	public Object[][] oneToManyTestCaseDataProvider()
-	{
-		return new Object[][] { { 1 }, { 2 }, { 3 }, { 4 } };
-	}
-
-	private void importBooksThenAuthors(OneToManyTestHarness.AuthorsAndBooks authorsAndBooks)
-	{
-		runAsSystem(() ->
-		{
-			dataService.add(authorsAndBooks.getBookMetaData().getName(), authorsAndBooks.getBooks().stream());
-			dataService.add(authorsAndBooks.getAuthorMetaData().getName(), authorsAndBooks.getAuthors().stream());
-			waitForIndexToBeStable(authorsAndBooks.getAuthorMetaData().getName(), indexService, LOG);
-			waitForIndexToBeStable(authorsAndBooks.getBookMetaData().getName(), indexService, LOG);
-		});
-	}
-
-	private void importAuthorsThenBooks(OneToManyTestHarness.AuthorsAndBooks authorsAndBooks)
-	{
-		runAsSystem(() ->
-		{
-			dataService.add(authorsAndBooks.getAuthorMetaData().getName(), authorsAndBooks.getAuthors().stream());
-			dataService.add(authorsAndBooks.getBookMetaData().getName(), authorsAndBooks.getBooks().stream());
-			waitForIndexToBeStable(authorsAndBooks.getAuthorMetaData().getName(), indexService, LOG);
-			waitForIndexToBeStable(authorsAndBooks.getBookMetaData().getName(), indexService, LOG);
-		});
-	}
-
-	private void deleteBooksThenAuthors(int testCase)
-	{
-		dataService.deleteAll("sys_Book" + testCase);
-		dataService.deleteAll("sys_Author" + testCase);
-	}
-
-	private void deleteAuthorsThenBooks(int testCase)
-	{
-		dataService.deleteAll("sys_Author" + testCase);
-		dataService.deleteAll("sys_Book" + testCase);
-	}
-
-	private OneToManyTestHarness.AuthorsAndBooks importAuthorsAndBooks(int testCase)
-	{
-		OneToManyTestHarness.AuthorsAndBooks authorsAndBooks;
-		switch (testCase)
-		{
-			case 1:
-				// case 1: books/authors both nillable, order of import not important
-				authorsAndBooks = oneToManyTestHarness.createAuthorAndBookEntities(1);
-				importBooksThenAuthors(authorsAndBooks);
-				return authorsAndBooks;
-			case 2:
-				// case 2: book.author required so add Author entities first
-				authorsAndBooks = oneToManyTestHarness.createAuthorAndBookEntities(2);
-				importAuthorsThenBooks(authorsAndBooks);
-				return authorsAndBooks;
-			case 3:
-				// case 3: author.books required so add Book entities first
-				authorsAndBooks = oneToManyTestHarness.createAuthorAndBookEntities(3);
-				importBooksThenAuthors(authorsAndBooks);
-				return authorsAndBooks;
-			case 4:
-				// case 4: books/authors both required
-				authorsAndBooks = oneToManyTestHarness.createAuthorAndBookEntities(4);
-				importBooksThenAuthors(authorsAndBooks);
-				return authorsAndBooks;
-			case 5:
-				// case 5: books/authors both nillable, ascending order
-				authorsAndBooks = oneToManyTestHarness.createAuthorAndBookEntities(5);
-				importBooksThenAuthors(authorsAndBooks);
-				return authorsAndBooks;
-			case 6:
-				// case 6: books/authors both nillable, descending order
-				authorsAndBooks = oneToManyTestHarness.createAuthorAndBookEntities(6);
-				importBooksThenAuthors(authorsAndBooks);
-				return authorsAndBooks;
-			default:
-				return null;
-		}
-	}
-
-	private List<Entity> importPersons(int testCase)
-	{
-		List<Entity> persons = oneToManyTestHarness.createPersonEntities(testCase);
-		runAsSystem(() ->
-		{
-			dataService.add(persons.get(0).getEntityMetaData().getName(), persons.stream());
-			waitForIndexToBeStable(persons.get(0).getEntityMetaData().getName(), indexService, LOG);
-		});
-		return persons;
-	}
-
-	@Test(singleThreaded = true, groups = GROUPS_ONE_TO_MANY)
-	@Transactional
-	public void testL1OneToManySingleEntityUpdate()
-	{
-		OneToManyTestHarness.AuthorsAndBooks authorsAndBooks = importAuthorsAndBooks(1);
-		try
-		{
-			Entity book1 = dataService.findOneById(authorsAndBooks.getBookMetaData().getName(), BOOK_1);
-			Entity author1 = dataService.findOneById(authorsAndBooks.getAuthorMetaData().getName(), AUTHOR_1);
-			Entity author2 = dataService.findOneById(authorsAndBooks.getAuthorMetaData().getName(), AUTHOR_2);
-
-			book1.set(BookMetaData1.AUTHOR, author2);
-			dataService.update(book1.getEntityMetaData().getName(), book1);
-
-			Entity author1RetrievedAgain = dataService
-					.findOneById(authorsAndBooks.getAuthorMetaData().getName(), author1.getIdValue());
-			assertEquals(Collections.emptyList(),
-					Lists.newArrayList(author1RetrievedAgain.getEntities(AuthorMetaData1.ATTR_BOOKS)));
-
-			Entity author2Retrieved = dataService
-					.findOneById(authorsAndBooks.getAuthorMetaData().getName(), author2.getIdValue());
-			Iterable<Entity> author2Books = author2Retrieved.getEntities(AuthorMetaData1.ATTR_BOOKS);
-
-			// expected behavior: book.author changed, new author.books order is undefined
-			Set<Object> retrievedAuthor2BookIds = StreamSupport.stream(author2Books.spliterator(), false)
-					.map(Entity::getIdValue).collect(toSet());
-			assertEquals(retrievedAuthor2BookIds, newHashSet(BOOK_2, BOOK_1));
-		}
-		finally
-		{
-			dataService.deleteAll(authorsAndBooks.getBookMetaData().getName());
-			dataService.deleteAll(authorsAndBooks.getAuthorMetaData().getName());
-		}
-	}
-
-	@Test(singleThreaded = true, groups = GROUPS_ONE_TO_MANY)
-	@Transactional
-	public void testL1OneToManyStreamingEntityUpdate()
-	{
-		OneToManyTestHarness.AuthorsAndBooks authorsAndBooks = importAuthorsAndBooks(1);
-		try
-		{
-			Entity book1 = dataService.findOneById(authorsAndBooks.getBookMetaData().getName(), BOOK_1);
-			Entity author1 = dataService.findOneById(authorsAndBooks.getAuthorMetaData().getName(), AUTHOR_1);
-			Entity author2 = dataService.findOneById(authorsAndBooks.getAuthorMetaData().getName(), AUTHOR_2);
-
-			book1.set(BookMetaData1.AUTHOR, author2);
-			dataService.update(book1.getEntityMetaData().getName(), Stream.of(book1));
-
-			Entity author1RetrievedAgain = dataService
-					.findOneById(authorsAndBooks.getAuthorMetaData().getName(), author1.getIdValue());
-			assertEquals(Collections.emptyList(),
-					Lists.newArrayList(author1RetrievedAgain.getEntities(AuthorMetaData1.ATTR_BOOKS)));
-
-			Entity author2Retrieved = dataService
-					.findOneById(authorsAndBooks.getAuthorMetaData().getName(), author2.getIdValue());
-			Iterable<Entity> author2Books = author2Retrieved.getEntities(AuthorMetaData1.ATTR_BOOKS);
-
-			// expected behavior: book.author changed, new author.books order is undefined
-			Set<Object> retrievedAuthor2BookIds = StreamSupport.stream(author2Books.spliterator(), false)
-					.map(Entity::getIdValue).collect(toSet());
-			assertEquals(retrievedAuthor2BookIds, newHashSet(BOOK_2, BOOK_1));
-		}
-		finally
-		{
-			dataService.deleteAll(authorsAndBooks.getBookMetaData().getName());
-			dataService.deleteAll(authorsAndBooks.getAuthorMetaData().getName());
-		}
-	}
-
-	@Test(singleThreaded = true, groups = GROUPS_ONE_TO_MANY)
-	@Transactional
-	public void testL1OneToManyEntitySingleEntityDelete()
-	{
-		OneToManyTestHarness.AuthorsAndBooks authorsAndBooks = importAuthorsAndBooks(1);
-		Entity book1 = dataService.findOneById(authorsAndBooks.getBookMetaData().getName(), BOOK_1);
-		Entity author1 = dataService.findOneById(authorsAndBooks.getAuthorMetaData().getName(), AUTHOR_1);
-
-		dataService.delete(book1.getEntityMetaData().getName(), book1);
-
-		Entity author1RetrievedAgain = dataService
-				.findOneById(authorsAndBooks.getAuthorMetaData().getName(), author1.getIdValue());
-		assertEquals(Collections.emptyList(),
-				Lists.newArrayList(author1RetrievedAgain.getEntities(AuthorMetaData1.ATTR_BOOKS)));
-
-		dataService.deleteAll(authorsAndBooks.getBookMetaData().getName());
-		dataService.deleteAll(authorsAndBooks.getAuthorMetaData().getName());
-	}
-
-	@Test(singleThreaded = true, groups = GROUPS_ONE_TO_MANY)
-	@Transactional
-	public void testL1OneToManyEntityStreamingEntityDelete()
-	{
-		OneToManyTestHarness.AuthorsAndBooks authorsAndBooks = importAuthorsAndBooks(1);
-		try
-		{
-			Entity book1 = dataService.findOneById(authorsAndBooks.getBookMetaData().getName(), BOOK_1);
-			Entity author1 = dataService.findOneById(authorsAndBooks.getAuthorMetaData().getName(), AUTHOR_1);
-
-			dataService.delete(book1.getEntityMetaData().getName(), Stream.of(book1));
-
-			Entity author1RetrievedAgain = dataService
-					.findOneById(authorsAndBooks.getAuthorMetaData().getName(), author1.getIdValue());
-			assertEquals(Collections.emptyList(),
-					Lists.newArrayList(author1RetrievedAgain.getEntities(AuthorMetaData1.ATTR_BOOKS)));
-		}
-		finally
-		{
-			dataService.deleteAll(authorsAndBooks.getBookMetaData().getName());
-			dataService.deleteAll(authorsAndBooks.getAuthorMetaData().getName());
-		}
-	}
-
-	@Test(singleThreaded = true, expectedExceptions = MolgenisDataException.class, groups = GROUPS_ONE_TO_MANY)
-	public void testOneToManyAuthorRequiredSetBooksNull()
-	{
-		// FIXME doesn't throw exception
-		OneToManyTestHarness.AuthorsAndBooks authorsAndBooks = importAuthorsAndBooks(2); // book.author required
-		Entity author = authorsAndBooks.getAuthors().get(0);
-		author.set(ATTR_BOOKS, null);
-		dataService.update(authorsAndBooks.getAuthorMetaData().getName(), author);
-	}
-
-	@Test(singleThreaded = true, expectedExceptions = MolgenisDataException.class, groups = GROUPS_ONE_TO_MANY)
-	public void testOneToManyAuthorRequiredSetAuthorsNull()
-	{
-		OneToManyTestHarness.AuthorsAndBooks authorsAndBooks = importAuthorsAndBooks(2); // book.author required
-		Entity book = authorsAndBooks.getBooks().get(0);
-		book.set(ATTR_AUTHOR, null);
-		dataService.update(authorsAndBooks.getBookMetaData().getName(), book);
-	}
-
-	@Test(singleThreaded = true, groups = GROUPS_ONE_TO_MANY)
-	public void testOneToManyAuthorRequiredUpdateValue()
-	{
-		OneToManyTestHarness.AuthorsAndBooks authorsAndBooks = importAuthorsAndBooks(2); // book.author required
-		String bookName = authorsAndBooks.getBookMetaData().getName();
-		String authorName = authorsAndBooks.getAuthorMetaData().getName();
-
-		Entity book = dataService.findOneById(bookName, BOOK_1);
-		book.set(ATTR_AUTHOR, dataService.findOneById(authorName, AUTHOR_2));
-		dataService.update(bookName, book);
-
-		Entity updatedBook = dataService.findOneById(bookName, BOOK_1);
-		assertEquals(updatedBook.getEntity(ATTR_AUTHOR).getIdValue(), AUTHOR_2);
-
-		// expected behavior: book.author changed, new author.books order is undefined
-		Entity updatedAuthor1 = dataService.findOneById(authorName, AUTHOR_1);
-		assertEquals(StreamSupport.stream(updatedAuthor1.getEntities(ATTR_BOOKS).spliterator(), false)
-				.map(Entity::getIdValue).collect(toSet()), newHashSet());
-
-		Entity updatedAuthor2 = dataService.findOneById(authorName, AUTHOR_2);
-		assertEquals(StreamSupport.stream(updatedAuthor2.getEntities(ATTR_BOOKS).spliterator(), false)
-				.map(Entity::getIdValue).collect(toSet()), newHashSet(BOOK_2, BOOK_1));
-	}
-
-	@Test(singleThreaded = true, expectedExceptions = MolgenisDataException.class, groups = GROUPS_ONE_TO_MANY)
-	public void testOneToManyBooksRequiredSetAuthorNull()
-	{
-		// FIXME doesn't throw exception
-		OneToManyTestHarness.AuthorsAndBooks authorsAndBooks = importAuthorsAndBooks(3); // author.books required
-		String bookName = authorsAndBooks.getBookMetaData().getName();
-
-		Entity book = dataService.findOneById(bookName, BOOK_1);
-		book.set(ATTR_AUTHOR, null);
-		dataService.update(bookName, book);
-	}
-
-	@Test(singleThreaded = true, expectedExceptions = MolgenisDataException.class, groups = GROUPS_ONE_TO_MANY)
-	public void testOneToManyBooksRequiredSetBooksNull()
-	{
-		OneToManyTestHarness.AuthorsAndBooks authorsAndBooks = importAuthorsAndBooks(3); // author.books required
-		String authorName = authorsAndBooks.getAuthorMetaData().getName();
-
-		Entity author = dataService.findOneById(authorName, AUTHOR_1);
-		author.set(ATTR_BOOKS, null);
-		dataService.update(authorName, author);
-	}
-
-	@Test(singleThreaded = true, groups = GROUPS_ONE_TO_MANY)
-	public void testOneToManyBookRequiredUpdateValue()
-	{
-		OneToManyTestHarness.AuthorsAndBooks authorsAndBooks = importAuthorsAndBooks(3); // book.author required
-		String bookName = authorsAndBooks.getBookMetaData().getName();
-		String authorName = authorsAndBooks.getAuthorMetaData().getName();
-
-		Entity author1 = dataService.findOneById(authorName, AUTHOR_1);
-		author1.set(ATTR_BOOKS, newArrayList(dataService.findOneById(bookName, BOOK_2)));
-		Entity author2 = dataService.findOneById(authorName, AUTHOR_2);
-		author2.set(ATTR_BOOKS, newArrayList(dataService.findOneById(bookName, BOOK_1)));
-		dataService.update(authorName, Stream.of(author1, author2));
-
-		Entity updatedAuthor = dataService.findOneById(authorName, AUTHOR_1);
-		assertEquals(updatedAuthor.getEntities(ATTR_BOOKS).iterator().next().getIdValue(), BOOK_2);
-
-		Entity updatedBook1 = dataService.findOneById(bookName, BOOK_1);
-		assertEquals(updatedBook1.getEntity(ATTR_AUTHOR).getIdValue(), AUTHOR_2);
-
-		Entity updatedBook2 = dataService.findOneById(bookName, BOOK_2);
-		assertEquals(updatedBook2.getEntity(ATTR_AUTHOR).getIdValue(), AUTHOR_1);
-	}
-
-	@Test(singleThreaded = true, expectedExceptions = MolgenisDataException.class, groups = GROUPS_ONE_TO_MANY)
-	public void testOneToManyBookAndAuthorRequiredSetAuthorNull()
-	{
-		OneToManyTestHarness.AuthorsAndBooks authorsAndBooks = importAuthorsAndBooks(4);
-		String bookName = authorsAndBooks.getBookMetaData().getName();
-
-		Entity book = dataService.findOneById(bookName, BOOK_1);
-		book.set(ATTR_AUTHOR, null);
-		dataService.update(bookName, book);
-	}
-
-	@Test(singleThreaded = true, expectedExceptions = MolgenisDataException.class, groups = GROUPS_ONE_TO_MANY)
-	public void testOneToManyBookAndAuthorRequiredSetBooksNull()
-	{
-		OneToManyTestHarness.AuthorsAndBooks authorsAndBooks = importAuthorsAndBooks(4);
-		String authorName = authorsAndBooks.getAuthorMetaData().getName();
-
-		Entity author = dataService.findOneById(authorName, AUTHOR_1);
-		author.set(ATTR_BOOKS, null);
-		dataService.update(authorName, author);
-	}
-
-	@Test(singleThreaded = true, groups = GROUPS_ONE_TO_MANY)
-	public void testOneToManyBookAndAuthorRequiredUpdateValue()
-	{
-		OneToManyTestHarness.AuthorsAndBooks authorsAndBooks = importAuthorsAndBooks(4);
-		String bookName = authorsAndBooks.getBookMetaData().getName();
-		String authorName = authorsAndBooks.getAuthorMetaData().getName();
-
-		Entity author1 = dataService.findOneById(authorName, AUTHOR_1);
-		Entity author2 = dataService.findOneById(authorName, AUTHOR_2);
-		Entity book1 = dataService.findOneById(bookName, BOOK_1);
-		Entity book2 = dataService.findOneById(bookName, BOOK_2);
-		book1.set(ATTR_AUTHOR, author2); // switch authors
-		book2.set(ATTR_AUTHOR, author1);
-
-		dataService.update(bookName, Stream.of(book1, book2));
-
-		assertEquals(dataService.findOneById(bookName, BOOK_1).getEntity(ATTR_AUTHOR).getIdValue(), AUTHOR_2);
-		assertEquals(dataService.findOneById(bookName, BOOK_2).getEntity(ATTR_AUTHOR).getIdValue(), AUTHOR_1);
-
-		Entity updatedAuthor1 = dataService.findOneById(authorName, AUTHOR_1);
-		assertEquals(StreamSupport.stream(updatedAuthor1.getEntities(ATTR_BOOKS).spliterator(), false)
-				.map(Entity::getIdValue).collect(toSet()), newHashSet(BOOK_2));
-
-		Entity updatedAuthor2 = dataService.findOneById(authorName, AUTHOR_2);
-		assertEquals(StreamSupport.stream(updatedAuthor2.getEntities(ATTR_BOOKS).spliterator(), false)
-				.map(Entity::getIdValue).collect(toSet()), newHashSet(BOOK_1));
 	}
 
 	@Test(singleThreaded = true)
